% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{S3_gee}
\alias{S3_gee}
\alias{.pval.summary.gee}
\alias{.pval.summary.geeglm}
\alias{confint_.gee}
\alias{nobsText.gee}
\alias{nobsText.geeglm}
\alias{nobsText.geese}
\alias{desc_.gee}
\title{Additional S3 methods for \link[gee]{gee} and \link[geepack]{geeglm} Objects}
\usage{
.pval.summary.gee(x)

.pval.summary.geeglm(x)

confint_.gee(x, level = 0.95, ...)

nobsText.gee(x)

nobsText.geeglm(x)

nobsText.geese(x)

desc_.gee(x, ...)
}
\arguments{
\item{x}{a \link[gee]{gee} or \link[geepack]{geeglm} object}

\item{level}{..}

\item{...}{additional parameters, currently of no use}
}
\description{
Additional S3 methods for \link[gee]{gee} and \link[geepack]{geeglm} Objects
}
\examples{
library(gee) # ?gee::gee
m1 = gee(breaks ~ tension, id=wool, data = warpbreaks, corstr="exchangeable")
# gee::gee object is badly defined; as well as package \pkg{gee}

library(geepack) # ?geepack::geeglm
m2 = geeglm(breaks ~ tension, id=wool, data = warpbreaks, corstr="exchangeable")
# I prefer using geepack::geeglm

}
