% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{vcov.gee}
\alias{vcov.gee}
\title{Variance-Covariance for \link[gee]{gee} Object}
\usage{
\method{vcov}{gee}(object, ...)
}
\arguments{
\item{object}{a \link[gee]{gee} object}

\item{...}{additional parameters, currently not in use}
}
\description{
..
}
\note{
Function \code{\link[=vcov.gee]{vcov.gee()}} is super important for \code{lmtest:::coeftest.glm}.
}
\seealso{
\code{geepack:::vcov.geeglm}.
}
